refactor: TodoInput 분리
